requirements.txt
torch>=2.11.0
torchvision>=0.26.0
rfdetr
opencv-python
pillow
numpy
supervision
pycocotools